feat: Creates PrioritizedIdentifierListenerProvider
The PrioritizedIdentifierListenerProvider mimics functionality present in the SharedEventManager (and implements the SharedEventManagerInterface).
Its purpose is to be a drop-in replacement for the `SharedEventManager` to allow users to start migrating to PSR-14 functionality.

In the process of working on this implementation, I discovered some complexity in the data structure returned from `getListenersForEventByPriority` implementation of `PrioritizedListenerProvider` that, when mimiced in `PrioritizedIdentifierListenerProvider`, made verifying behavior difficult.
In particular, it was this line:

```php
$prioritizedListeners[$priority][] = $listOfListeners[0];
```

The problem that arose is that the `$prioritizedListeners` returned were now two levels deep, which made comparisons far harder. I changed this to read:

```
$prioritizedListeners[$priority] = isset($prioritizedListeners[$priority])
    ? array_merge($prioritizedListeners[$priority], $listOfListeners[0])
    : $listOfListeners[0];
```

This makes the return value far simpler, and _should_ keep speed reasonable, though I have yet to benchmark it.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>