feat: Creates a PrioritizedAggregateListenerProvider
This version acts like the combination of EventManager+SharedEventManager in terms of how it aggregates and resolves priority for listeners.

The class aggregates a list of `PrioritizedListenerAttachmentInterface` instances (and implements the interface itself), looping over each in ordert to build up a prioritized list of all listeners from all providers.
Since they are done in order, the order in which they should be attached generally is:

- PrioritizedListenerProvider
- PrioritizedIdentifierListenerProvider

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>